refactor: move shared types to customer.types.ts and update imports
