+++ src/api/signup/createCustomer.types.ts
@@ -1,31 +1,10 @@
-export enum CountryCode {
-    FR = 'FR',
-    DE = 'DE',
-    IT = 'IT',
-}
+import type { BaseCustomer } from '../../types/customer.types';
 
-interface Address {
-    id?: string; // generated by the server
-    key?: string; // min: 2, max: 256, ^[A-Za-z0-9_-]+$
-    country: CountryCode;
-    streetName: string;
-    postalCode: string;
-    city: string;
-}
-
-export interface Customer {
-    key?: string; // min: 2, max: 256, ^[A-Za-z0-9_-]+$
-    customerNumber?: string;
-    email: string;
+export interface Customer extends BaseCustomer {
     password: string;
-    firstName: string;
-    lastName: string;
-    dateOfBirth: string;
-    addresses: Address[];
+    customerNumber?: string;
     shippingAddresses: number[];
     billingAddresses: number[];
-    defaultShippingAddress?: number;
-    defaultBillingAddress?: number;
 }
 
 export interface CreateCustomerResponse {
