+++ src/components/blocks/UserProfile/parts/AddressList/AddressList.tsx
@@ -5,8 +5,8 @@ import { Input } from '../../../../../components/ui/Input';
 import { Label } from '../../../../../components/ui/Label';
 import { Select } from '../../../../../components/ui/Select';
 import { ConfirmModal } from '../../../../../components/ui/ConfirmModal';
-import { CountryCode } from '../../../../../api/signup/createCustomer.types';
-import type { Address } from '../../../../../api/profile/profile.types';
+import { CountryCode } from '../../../../../types/customer.types';
+import type { Address } from '../../../../../types/customer.types';
 import { FiEdit2, FiTrash2, FiPlus } from 'react-icons/fi';
 import styles from './AddressList.module.scss';
 import { useAddressList } from '../../lib/address-hooks/useAddressList';
